Verify that both ipv4 and ipv6 are available in port selection logic (#17)
Slightly funky but currently the `net.Listen` call seems to only check whether the ipv6 host is available (when using `:port` as addr) rather than both ipv4 and ipv6. This break with the tilt setup as supabase storage api is listening on port 5000 on ipv4. 

We then listen on the ipv4 addr, instead of the ipv6 addr, as for some reason my browser won't resolve to the ipv6 addr.