In 'project init', use the project human ID as the technical project name. (#67)
The detection from Unity's `productName` was unreliable, whereas the human ID always has the same reliable format.

The solution name is now always just `Server.sln`. It was mainly named uniquely so that the different servers show up differently in Visual Studio's menu, but this is not a big concern with new projects.