Fix handling of `environments: null` in `metaplay-project.yaml` (#88)
This fixes the `metaplay update project-environments` to work correctly with the implicit `environments: null`. The implicit null is generated when initializing the SDK in a project without any environments present.

Revert the generation of `environments: []` for projects with no envs (in portal). It causes more problems than it fixes. It was never published so no need to deal with that syntax in `update proj-envs`.